feat: adding select widget
