+++ catalyst_voices/apps/voices/lib/pages/proposal_builder/proposal_builder_rich_text_step.dart
@@ -60,9 +60,9 @@ class _ProposalBuilderRichTextStepState
     return SectionStepStateBuilder(
       id: widget.step.sectionStepId,
       builder: (context, value, child) {
-        return WorkspaceTileContainer(
+        return SelectableTile(
           isSelected: value.isSelected,
-          content: child!,
+          child: child!,
         );
       },
       child: VoicesRichText(
