+++ catalyst_voices/apps/voices/lib/pages/treasury/steps/treasury_campaign_stages_edit_step.dart
@@ -20,9 +20,9 @@ class TreasuryCampaignStagesEditStep extends StatelessWidget {
     return SectionStepStateBuilder(
       id: step.sectionStepId,
       builder: (context, value, child) {
-        return WorkspaceTileContainer(
+        return SelectableTile(
           isSelected: value.isSelected,
-          content: Column(
+          child: Column(
             children: [
               TreasuryCampaignStepHeader(step: step),
               const SizedBox(height: 12),
